create separate package for the github fetcher
